Fold the KDA track-state test back into test_kda_prefill.py
The taxonomy check only admits new registered files under
test/registered/<kind>/<subsystem>/ and kernel-kind files require a
*-kernel-* suite, which does not exist for gb300 - so the standalone file
at kernel/ops/attention/ had to drop its gb300 registration. Extending the
existing registered sibling (M files are exempt from the taxonomy check)
keeps the legacy location next to the other chunk_kda tests and restores
the b200 + gb300 coverage.
